I might as well use doubles

+++ weighting.cpp
@@ -16,12 +16,12 @@ const int WEIGHT_RESOLUTION = 512;
 // sphere. Assumes 90 degree field of view.
 //
 // TODO: This doesn't include cos(angle from normal) factor.
-void projWeights(std::vector<GLfloat> &weights)
+void projWeights(std::vector<GLdouble> &weights)
 {
     // From -1 to +1.
-    GLfloat conv = 2.0 / WEIGHT_RESOLUTION;
+    GLdouble conv = 2.0 / WEIGHT_RESOLUTION;
     // Make total come to 1.
-    GLfloat weight = 3.0 / (M_PI * 2.0);
+    GLdouble weight = 3.0 / (M_PI * 2.0);
 
     for (int y = 0; y < WEIGHT_RESOLUTION; ++y) {
         for (int x = 0; x < WEIGHT_RESOLUTION; ++x) {
@@ -37,17 +37,17 @@ void projWeights(std::vector<GLfloat> &weights)
 // Like projWeights, but axis-aligns things first.
 //
 // First step towards doing an analytic version.
-void projWeights2(std::vector<GLfloat> &weights)
+void projWeights2(std::vector<GLdouble> &weights)
 {
     // From -1 to +1.
-    GLfloat conv = 2.0 / WEIGHT_RESOLUTION;
+    GLdouble conv = 2.0 / WEIGHT_RESOLUTION;
     // Make total come to 1.
-    GLfloat weight = 3.0 / (M_PI * 2.0);
+    GLdouble weight = 3.0 / (M_PI * 2.0);
 
     for (int y = 0; y < WEIGHT_RESOLUTION; ++y) {
         for (int x = 0; x < WEIGHT_RESOLUTION; ++x) {
-            GLfloat px = (x + 0.5) * conv - 1.0;
-            GLfloat py = (y + 0.5) * conv - 1.0;
+            GLdouble px = (x + 0.5) * conv - 1.0;
+            GLdouble py = (y + 0.5) * conv - 1.0;
             double dist = sqrt(px * px + py * py);
 
             Vertex v1 = Vertex(dist - conv/2, -0.5 * conv, 1);
@@ -64,10 +64,10 @@ void projWeights2(std::vector<GLfloat> &weights)
 // TODO: I should really get some cpp_unit, shouldn't I?
 int main(void)
 {
-    std::vector<GLfloat> ws, ws2;
+    std::vector<GLdouble> ws, ws2;
     projWeights(ws);
     projWeights2(ws2);
-    GLfloat total = 0.0, total2 = 0.0, maxError = 0.0;
+    GLdouble total = 0.0, total2 = 0.0, maxError = 0.0;
     for (int i = 0, n = ws.size(); i < n; ++i) {
         // if (i % WEIGHT_RESOLUTION == 0) std::cout << std::endl;
         // std::cout << ws[i] << " vs " << ws2[i] << "(" << ws[i] / ws2[i] << ")" << std::endl;
